requirements.txt
django==3.2.5
djangorestframework==3.12.4
drf-yasg==1.20.0
psycopg2-binary==2.9.1
gunicorn==20.1.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-bootstrap3==15.0.0